+++ app/src/main/kotlin/com/prajwalch/torrentsearch/ui/home/component/RecentSearchesCard.kt
@@ -3,14 +3,17 @@ package com.prajwalch.torrentsearch.ui.home.component
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
 import androidx.compose.material3.ListItem
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedCard
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
@@ -23,21 +26,34 @@ import com.prajwalch.torrentsearch.ui.theme.spaces
 fun RecentSearchesCard(
     queries: List<String>,
     onQueryClick: (String) -> Unit,
+    onClose: () -> Unit,
     modifier: Modifier = Modifier,
 ) {
     OutlinedCard(modifier = modifier) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(top = MaterialTheme.spaces.large),
-            verticalArrangement = Arrangement.spacedBy(MaterialTheme.spaces.small),
-        ) {
-            Text(
-                modifier = Modifier.padding(horizontal = MaterialTheme.spaces.large),
-                text = stringResource(R.string.home_title_recent_searches),
-                color = MaterialTheme.colorScheme.primary,
-                style = MaterialTheme.typography.titleMedium,
-            )
+        Column(modifier = Modifier.fillMaxWidth()) {
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(
+                        horizontal = MaterialTheme.spaces.large,
+                        vertical = MaterialTheme.spaces.extraSmall,
+                    ),
+                horizontalArrangement = Arrangement.SpaceBetween,
+                verticalAlignment = Alignment.CenterVertically,
+            ) {
+                Text(
+                    text = stringResource(R.string.home_title_recent_searches),
+                    color = MaterialTheme.colorScheme.primary,
+                    style = MaterialTheme.typography.titleMedium,
+                )
+
+                IconButton(onClick = onClose) {
+                    Icon(
+                        painter = painterResource(R.drawable.ic_close),
+                        contentDescription = null,
+                    )
+                }
+            }
 
             Column {
                 queries.forEach { query ->